backend/requirements.txt
# WatchTower Backend - FastAPI
# AI-Powered Cyber Incident Portal for Defence

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
supabase>=2.0.0
httpx>=0.24.0

# AI/ML for Threat Analysis
google-generativeai==0.3.2
openai==1.12.0

# URL Analysis
tldextract==5.1.1
validators==0.22.0
dnspython==2.5.0
whois==0.9.27

# File Analysis
python-magic-bin==0.4.14

# Security
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# CORS & Utils
aiofiles==23.2.1
jinja2==3.1.3

# Rate Limiting
slowapi==0.1.9

# Environment
python-dotenv==1.0.1
